Safari-609.1.11.1/LayoutTests/js/dom/promise-rejection-event-should-follow-webidl-promise-conversion-rule-expected.txt

Test PromiseRejectionEvent constructor follows WebIDL promise conversion rule.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS event.promise is an instance of Promise
PASS event.reason is "Cappuccino"
PASS value is 42
PASS event.promise is an instance of Promise
PASS event.promise is promise
PASS event.reason is "Cocoa"
PASS successfullyParsed is true

TEST COMPLETE
